Tests/Challenges/ArraysAndStrings/TwoPointers/FindMaxAverageTests.cs
using Challenges.ArraysAndStrings.TwoPointers;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Tests.Challenges.ArraysAndStrings.TwoPointers
{
    public class FindMaxAverageTests
    {
        private FindMaxAverage _underTests;

        public FindMaxAverageTests()
        {
            _underTests = new FindMaxAverage();
        }


        [Test]
        [TestCase(new int[] { 1, 12, -5, -6, 50, 3 }, 4, 12.75000)]
        [TestCase(new int[] { 5 }, 1, 5)]
        [TestCase(new int[] { -1 }, 1, -1)]
        [TestCase(new int[] { 4433, -7832, -5068, 4009, 2830, 6544, -6119, -7126, -780, -4254, -8249, -9168, 9492, 402, 5789, 6808, 8953, 5810, -7353, 7933, 4766, 5182, -3230, -1989, 5786, 6922, -4646, 4415, -9906, 807, -6373, 3370, 2604, 8751, -9173, -2668, -6876, 9500, 3465, -1900, 4134, -1758, -1453, -5201, -9825, 4469, -1999, -1108, 1836, 3923, 6796, -5252, 9863, -5997, -3251, 9596, -3404, -540, 2826, -1737, 3341, -3623, -9885, 2603, -5782, 8174, 2710, 6504, -4128 }, 59, 526.37288)]
        [TestCase(new int[] { 8860, -853, 6534, 4477, -4589, 8646, -6155, -5577, -1656, -5779, -2619, -8604, -1358, -8009, 4983, 7063, 3104, -1560, 4080, 2763, 5616, -2375, 2848, 1394, -7173, -5225, -8244, -809, 8025, -4072, -4391, -9579, 1407, 6700, 2421, -6685, 5481, -1732, -8892, -6645, 3077, 3287, -4149, 8701, -4393, -9070, -1777, 2237, -3253, -506, -4931, -7366, -8132, 5406, -6300, -275, -1908, 67, 3569, 1433, -7262, -437, 8303, 4498, -379, 3054, -6285, 4203, 6908, 4433, 3077, 2288, 9733, -8067, 3007, 9725, 9669, 1362, -2561, -4225, 5442, -9006, -429, 160, -9234, -4444, 3586, -5711, -9506, -79, -4418, -4348, -5891 }, 93, -594.58065)]

        public void FindMaxAverageTest(int[] nums, int k, double expected)
        {
            double result = _underTests.MaxAverage(nums, k);
            Assert.That(result, Is.EqualTo(expected));
        }
    }
}
